Fixed errors related to the theme
