+++ src/pages/loginPage/Login.css
@@ -21,7 +21,6 @@
   display: flex;
   flex-direction: column;
   gap: 20px;
-  background-color: #fff;
   border: none;
   border-radius: 8px;
   box-shadow: 0 2px 4px rgba(227, 211, 211, 0.837),
