+++ src/pages/loginPage/LoginPage.tsx
@@ -1,9 +1,9 @@
-import { Box, Typography } from "@mui/material";
+import { Box, Typography, useTheme } from "@mui/material";
 import { useState } from "react";
 import SwitchSelector from "react-switch-selector";
 import SignUpForm from "../../components/loginPageComponents/SignUpForm";
 import LoginForm from "../../components/loginPageComponents/LoginForm";
-import "./Login.scss";
+import "./Login.css";
 
 function LoginPage() {
   const [checked, setChecked] = useState(false);
@@ -28,6 +28,8 @@ function LoginPage() {
     ({ value }) => value === "logIn"
   );
 
+  const theme = useTheme();
+
   return (
     <>
       <Box className="container-login">
@@ -35,15 +37,17 @@ function LoginPage() {
           <Typography
             sx={{
               fontWeight: 800,
-              fontSize: "44px",
-              color: "primary.main",
               marginBottom: "10px",
             }}
+            variant="h3"
+            color="primary.main"
           >
             CinemaHub
           </Typography>
           <Typography
-            sx={{ fontWeight: 600, fontSize: "20px", fontStyle: "italic" }}
+            variant="h6"
+            color="text.primary"
+            sx={{ fontWeight: 600, fontStyle: "italic" }}
           >
             Your ticket to the world of entertainment!
           </Typography>
@@ -54,13 +58,13 @@ function LoginPage() {
               onChange={handleChange}
               options={options}
               initialSelectedIndex={initialSelectedIndex}
-              backgroundColor={"#cfd1d9"}
-              selectedBackgroundColor={"#0075e8"}
+              backgroundColor={theme.palette.secondary.main}
+              selectedBackgroundColor={theme.palette.primary.main}
               fontSize={20}
-              selectedColor={"white"}
+              fontColor={theme.palette.text.secondary}
+              selectedFontColor={theme.palette.text.primary}
             />
           </div>
-
           {checked ? <SignUpForm /> : <LoginForm />}
         </Box>
       </Box>
